changed /kick commands
